django/requirements.txt
Django==4.2.6
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
requests==2.31.0
bs4==0.0.1
openpyxl==3.1.2
pandas==2.1.4
networkx==3.2.1